+++ index.html
@@ -7,25 +7,25 @@
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lobster&effect=shadow-multiple">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="./style.css">
-    <title>Awesome books</title>
+    <title>Awesome-book-chester</title>
 </head>
 <body>
     <header>
-        <nav class="nav-bar wraper">
-            <div class="logo">Awesome Books</div>
-            <ul class="nav-bar-options">
-                <li class="nav-link list-link"><a class="link" href="#lists">List <span>|</span></a></li>
-                <li class="nav-link add-new-link"><a class="link active" href="#newbook">Add New <span>|</span></a></li>
-                <li class="nav-link contact-link"><a class="link" href="#contact">Contact </a></li>
+        <nav class="nav-bar wrap">
+            <div class="logo">Awesome Book</div>
+            <ul class="opt">
+                <li class="navigation list-link"><a class="link" href="#lists">List <span>|</span></a></li>
+                <li class="navigation add-link"><a class="link active" href="#newbook">Add New <span>|</span></a></li>
+                <li class="navigation contact-link"><a class="link" href="#contact">Contact </a></li>
             </ul>
         </nav>
-        <div class="date-time">
-            <p class="wraper" id="current-date"></p>
+        <div class="dateTIME">
+            <p class="wrap" id="current-date"></p>
         </div>
     </header>
     <div>
-        <div class="form-list">
-            <div class="book-list">
+        <div class="FORM-1">
+            <div class="list-books">
                 <h1>All awesome books</h1>
                 <table>
                     <tbody id="tbody">
@@ -34,9 +34,9 @@
                 </table>
             </div>
             
-            <div class="form-title-add">
-                <h2 class="title-form-add">Add a new book</h2>
-                <div class="form-container">
+            <div class="FORM-2">
+                <h2 class="FORM-3">Add a new book</h2>
+                <div class="FORM-4">
                     <form id="form">
                         <input type="text" name="title" id="title" placeholder="Title" required> 
                         <br><br>
@@ -47,13 +47,13 @@
                 </div>
             </div>
             <div class="contact hidden">
-                <h2 class="title-contact">Contact Information</h2>
+                <h2 class="title-1">Contact Information</h2>
                 <p>Do you have any questions or you just want to say "Hello"?</p>
                 <p class="contact-last-paragraph">You can reach out to us!</p>
                 <ul class="contact-details">
-                    <li class="contact-list">* Our email: mail@mail.com</li>
-                    <li class="contact-list">* Our phone number: 123456789</li>
-                    <li class="contact-list">* Our address: Streetname 123 City, Country</li>
+                    <li class="info">* Our email: mail@mail.com</li>
+                    <li class="info">* Our phone number: 123456789</li>
+                    <li class="info">* Our address: Streetname 123 City, Country</li>
                 </ul>
             </div>
         </div>
