+++ modules/books.js
@@ -1,0 +1,6 @@
+export default class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
